+++ lib/transpec/syntax/mixin/monkey_patch_any_instance.rb
@@ -55,12 +55,7 @@ module Transpec
           receiver_node, method_name = *subject_node
           return nil unless receiver_node
           return nil unless method_name == :any_instance
-
-          if receiver_node.const_type? || receiver_node == s(:send, nil, :described_class)
-            receiver_node
-          else
-            nil
-          end
+          receiver_node
         end
       end
     end